refactor: modules and states folder structure
