+++ apps/app/pages/[workspaceSlug]/projects/[projectId]/modules/index.tsx
@@ -12,7 +12,7 @@ import { requiredAuth } from "lib/auth";
 import projectService from "services/project.service";
 import modulesService from "services/modules.service";
 // components
-import SingleModuleCard from "components/project/modules/single-module-card";
+import { SingleModuleCard } from "components/modules";
 // ui
 import { EmptySpace, EmptySpaceItem, HeaderButton, Loader } from "components/ui";
 import { BreadcrumbItem, Breadcrumbs } from "components/breadcrumbs";
